changed int to float

+++ app/CoreContext/Users/Infrastructure/Actions/CreateArrayUserWorth.php
@@ -4,7 +4,7 @@ namespace App\CoreContext\Users\Infrastructure\Actions;
 
 class CreateArrayUserWorth
 {
-    public static function execute($userWallets, $companies): int
+    public static function execute($userWallets, $companies): float
     {
         $worth = 0;
         foreach ($userWallets as $wallet)
